backend/auth-service/requirements.txt
fastapi
uvicorn
asyncpg
PyJWT
bcrypt
httpx
pydantic[email]
python-dotenv
resend
